+++ changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- lnd-grpc-client: ability to specify cert and macaroon as base64-encoded value
 
 ### Changed
 - upgrade: update jmolecules bom from v2022.2.4 to v2022.3.0